Use compose destinations for handling compose screens
