+++ app/src/main/AndroidManifest.xml
@@ -16,12 +16,7 @@
         android:theme="@style/Theme.MovieApp3"
         tools:targetApi="31">
         <activity
-            android:name=".details.MovieDetailsActivity"
-            android:exported="false"
-            android:label="@string/title_activity_movie_details"
-            android:theme="@style/Theme.MovieApp3" />
-        <activity
-            android:name=".listing.presentation.MoviesActivity"
+            android:name=".common.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MovieApp3">
