+++ build.gradle.kts
@@ -1,6 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id("com.android.application") version "8.1.0" apply false
-    id("org.jetbrains.kotlin.android") version "1.8.10" apply false
+    id("com.google.devtools.ksp") version "1.8.20-1.0.10"
+    id("org.jetbrains.kotlin.android") version "1.8.20" apply false
     id ("com.google.dagger.hilt.android") version "2.44" apply false
 }
